refactor: extract settings runway workflow
